:sparkles::boom: Generalized the generation of selects for search
Any field that is displayed on the webpage can be used to create
options in a select element.
Example:
<select class="bibtex_search bibtex_generate_author" search="author">

This breaks the previous version of "bibtex_author". However, now any
field can be used as in:
<select class="bibtex_search bibtex_generate_year" search="year">
or
<select class="bibtex_search bibtex_generate_title" search="title">
with the understand the the information generated in the select's
options comes from the webpage and not the bibtex file. Therefore if
the user would like to generate a list of field but doesn't want that
information to be visible on the webpage they would need to hide it:
<span class="bibtexkey" style="display:none"></span>
<span class="bibtextypekey" style="display:none"></span>